Extend the no-op cast vein: drop 157 casts that restate their own operand's type
Two shapes, both gated per file on the .o md5 with automatic revert; every
object below is byte-identical, whole tree 99.75305 / 885 complete, main.dol OK.

Field/global loads (7 casts, 5 files): a cast on a FIELD or a scalar global
restates what the load already produced, so it is noise - unlike a cast on a
local, which forces a real extsh/extsb. The detector requires the name to carry
exactly one scalar declaration repo-wide and to have no pointer/aggregate
declaration anywhere, which keeps it off same-named fields of differing width.

Function returns (150 casts, 53 files): same argument one step out - a call
already delivers its declared return type in r3, so (u32)mainGetBit(...),
(u8)Obj_IsLoadingLocked(), (int)Obj_IsObjectAlive(...) and friends are no-ops.
Gated on a repo-unique non-void, non-pointer prototype.

Seven files failed the md5 gate and were reverted whole, so their casts are
genuinely load-bearing - the TU sees a return width that differs from the
repo-wide prototype: 700_Andross and 655_WCPressureS (mainGetBit),
488_SB_Galleon (sbGetPropeller), engine/67 (camcontrol_samplePathState),
engine/10_expgfx (Obj_IsLoadingLocked), main/vecmath (rand), main/shader
(SaveGame_findTransientMapBit).

Peer lanes and every file with a commit under two hours old were held out.

+++ src/dlls/engine/10_expgfx/expgfx.c
@@ -3079,7 +3079,7 @@ void expgfx_updateActivePools(u8 sourceMode, int sourceId, int resetSourceFrameS
                         axisX = 250.0f * (workA / norm);
                         axisY = 250.0f * (workB / norm);
                         axisZ = 250.0f * (attractRatio / norm);
-                        attractRatio = 2.0f / (EXPGFX_U16_TO_UNIT_SCALE * (f32)(u16)slot->scaleTarget);
+                        attractRatio = 2.0f / (EXPGFX_U16_TO_UNIT_SCALE * (f32)slot->scaleTarget);
                         quad[0].x = (s16)axisX;
                         quad[0].y = (s16)axisY;
                         quad[0].z = (s16)axisZ;
